Welcome page that works when hs_restclient is down
Tony and Dandong, you may have suggestions for making the code in the second block cleaner, but this approach could be used when your Welcome.ipynb is not functional because of hs_restclient issues.  This way users can still use their usual workflows that depend on JupyterHub, as long as they make their resources public.